worker/requirements.txt
boto3
redis
Pillow
prometheus-client
pytest
